Move some nursegui code to main_window
